backend/requirements/base.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
langchain
langchain-core
langchain-community
langchain-tavily
langgraph
langchain-groq
tavily-python
pymupdf